added exception handling

+++ server/hodl-invest/src/main/java/com/kenny/hodlinvest/controller/UserController.java
@@ -1,5 +1,7 @@
 package com.kenny.hodlinvest.controller;
 
+import com.kenny.hodlinvest.exception.UserException;
+import com.kenny.hodlinvest.exception.UserNotFoundException;
 import com.kenny.hodlinvest.model.User;
 import com.kenny.hodlinvest.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -35,7 +37,11 @@ public class UserController {
             path = "{username}"
     )
     public User getUserByName(@PathVariable String username){
-        return userService.getUserByName(username);
+        if(!userService.userExists(username))
+            throw new UserNotFoundException("User does not exist");
+
+            return userService.getUserByName(username);
+
     }
 
     @RequestMapping(
@@ -43,6 +49,9 @@ public class UserController {
             produces = MediaType.APPLICATION_JSON_VALUE
     )
     public void addNewUser(@RequestBody User user){
+        if(userService.userExists(user.getUsername())){
+            throw new UserException("Username already exists.");
+        }
         userService.addUser(user.getUsername(), user);
     }
 
@@ -51,6 +60,9 @@ public class UserController {
             path = "{username}"
     )
     public void deleteUserByName(@PathVariable String username){
+        if(!userService.userExists(username))
+            throw new UserNotFoundException("User does not exist");
+
         userService.deleteUserByName(username);
     }
 
@@ -59,6 +71,9 @@ public class UserController {
             path = "{username}/transactions/{amount}"
     )
     public void updateUserPlayMoney(@PathVariable String username, @PathVariable double amount){
+        if(!userService.userExists(username))
+            throw new UserNotFoundException("User does not exist");
+
         userService.updateUserPlayMoney(username, amount);
     }
 }
